feat(wasm): Add Emscripten/embind binding and make wasm targets
Wraps the Bundle 3.1 C ABI as a JavaScript-friendly Workbook class plus
an evalFormula convenience function via embind. The native build stays
unaffected (FM_BUILD_WASM defaults OFF; Emscripten is not required for
CI). Three Make targets land:
  * make wasm         -> build-wasm/formulon.{js,wasm} (Release)
  * make wasm-debug   -> build-wasm-debug/ (-O0 -g, assertions on)
  * make test-wasm    -> Node ESM smoke runner with no npm deps

embind.cpp is the only TU compiled with -frtti so embind's typeid-keyed
type registry works at runtime; the rest of the engine stays
-fno-rtti / -fno-exceptions. Failures cross the boundary as plain
JsStatus envelopes ({ ok, status, message, context }), not thrown JS
exceptions.

Initial size: formulon.wasm = 1.29 MiB raw / 342 KiB Brotli, well
inside the 1.8 MB / 600 KB ceiling. A POST_BUILD step prints the size
on every wasm rebuild so regressions surface immediately.

The single-line size_t cast in builtins/text.cpp::Rept silences a
-Wshorten-64-to-32 warning that only fires under wasm32 (where size_t
is 32-bit); the existing kExcelTextCapUnits bound check keeps it safe.

+++ src/eval/builtins/text.cpp
@@ -208,7 +208,11 @@ Value Rept(const Value* args, std::uint32_t /*arity*/, Arena& arena) {
     return Value::error(ErrorCode::Value);
   }
   std::string out;
-  out.reserve(text.value().size() * reps);
+  // The `unit_len * reps` overflow check above guarantees the product
+  // fits within Excel's text-cap and therefore within `size_t` on every
+  // supported target (including 32-bit wasm32). Cast explicitly to
+  // silence `-Wshorten-64-to-32` under wasm32 where `size_t` is 32-bit.
+  out.reserve(static_cast<std::size_t>(text.value().size() * reps));
   for (std::uint64_t i = 0; i < reps; ++i) {
     out.append(text.value());
   }
